+++ Client/src/client/ReportSubscriberUI.fxml
@@ -1,50 +1,88 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.chart.PieChart?>
+<?import javafx.scene.chart.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
 
-<AnchorPane xmlns="http://javafx.com/javafx/23.0.1" 
-            xmlns:fx="http://javafx.com/fxml"
-            styleClass="main-pane"
-            stylesheets="@ReportScreen.css"
-            fx:controller="client.ReportSubscriberController">
+<AnchorPane xmlns="http://javafx.com/javafx/23.0.1"
+	xmlns:fx="http://javafx.com/fxml"
+	fx:controller="client.ReportSubscriberController"
+	stylesheets="@ReportScreen.css" styleClass="main-pane">
 
-    <VBox spacing="20.0" alignment="CENTER"
-          AnchorPane.topAnchor="20.0"
-          AnchorPane.leftAnchor="20.0"
-          AnchorPane.rightAnchor="20.0"
-          AnchorPane.bottomAnchor="20.0"
-          styleClass="menu-container"
-          style="-fx-padding: 20;">
-        
-        <TableView fx:id="subscriberTable" prefHeight="400.0" prefWidth="960.0"
-                   styleClass="main-table">
-            <columns>
-                <TableColumn fx:id="colSubscriberId" text="Subscriber ID" prefWidth="120.0"/>
-                <TableColumn fx:id="colName" text="Name" prefWidth="140.0"/>
-                <TableColumn fx:id="colEmail" text="Email" prefWidth="180.0"/>
-                <TableColumn fx:id="colPhone" text="Phone" prefWidth="130.0"/>
-                <TableColumn fx:id="colTag" text="Tag" prefWidth="120.0"/>
-                <TableColumn fx:id="colRole" text="Role" prefWidth="100.0"/>
-            </columns>
-        </TableView>
-        
-        <PieChart fx:id="rolePieChart" prefHeight="240.0" prefWidth="400.0" />
+	<VBox spacing="20.0" alignment="CENTER"
+		AnchorPane.topAnchor="20.0" AnchorPane.leftAnchor="20.0"
+		AnchorPane.rightAnchor="20.0" AnchorPane.bottomAnchor="20.0"
+		style="-fx-padding: 20;" styleClass="menu-container">
 
-        <HBox spacing="15.0" alignment="CENTER">
-            <Label text="Role:" />
-            <ComboBox fx:id="roleComboBox" prefWidth="120"/>
-            <Button fx:id="sortBySubscriberIdButton" text="Sort by Subscriber ID" styleClass="button-blue"/>
-            <Button fx:id="exportCsvButton" text="Export as CSV" styleClass="button-blue"/>
-        </HBox>
+		<!-- TOP SECTION: Subscriber Table -->
+		<TableView fx:id="subscriberTable" prefHeight="300.0"
+			styleClass="main-table">
+			<columns>
+				<TableColumn fx:id="colSubscriberId"
+					text="Subscriber ID" prefWidth="120.0" />
+				<TableColumn fx:id="colName" text="Name"
+					prefWidth="120.0" />
+				<TableColumn fx:id="colEmail" text="Email"
+					prefWidth="180.0" />
+				<TableColumn fx:id="colPhone" text="Phone"
+					prefWidth="140.0" />
+				<TableColumn fx:id="colTag" text="Tag" prefWidth="140.0" />
+				<TableColumn fx:id="colRole" text="Role"
+					prefWidth="100.0" />
+			</columns>
+		</TableView>
 
-        <Button fx:id="backButton" text="Back" styleClass="button-blue" onAction="#handleBackButton"/>
-    </VBox>
+		<HBox spacing="15.0" alignment="CENTER">
+			<Label text="Role:" />
+			<ComboBox fx:id="roleComboBox" prefWidth="140.0" />
+			<Button fx:id="sortBySubscriberIdButton"
+				text="Sort by Subscriber ID" styleClass="button-blue" />
+			<Button fx:id="exportCsvButton" text="Export as CSV"
+				styleClass="button-blue" />
+			<Button fx:id="backButton" text="Back"
+				styleClass="button-blue" onAction="#handleBackButton" />
+		</HBox>
+
+		<!-- BOTTOM SECTION: History View -->
+		<SplitPane orientation="HORIZONTAL" dividerPositions="0.55"
+			prefHeight="330.0">
+			<!-- LEFT SIDE: History Table -->
+			<VBox spacing="10.0" alignment="TOP_LEFT"
+				style="-fx-padding: 10;">
+				<HBox spacing="10.0">
+					<TextField fx:id="subscriberIdField"
+						promptText="Subscriber ID" prefWidth="150.0" />
+					<Button fx:id="viewHistoryButton"
+						text="View Subscriber History" styleClass="button-blue" />
+				</HBox>
+
+				<TableView fx:id="historyTable" prefHeight="250.0">
+					<columns>
+						<TableColumn fx:id="colSessionId" text="Session ID"
+							prefWidth="80.0" />
+						<TableColumn fx:id="colSpotId" text="Spot ID"
+							prefWidth="80.0" />
+						<TableColumn fx:id="colInTime" text="In Time"
+							prefWidth="160.0" />
+						<TableColumn fx:id="colOutTime" text="Out Time"
+							prefWidth="160.0" />
+						<TableColumn fx:id="colLate" text="Late?"
+							prefWidth="60.0" />
+					</columns>
+				</TableView>
+			</VBox>
+
+			<!-- RIGHT SIDE: Bar Chart -->
+			<BarChart fx:id="barChart" title="Sessions Per Month"
+				prefHeight="300.0" prefWidth="400.0">
+				<xAxis>
+					<CategoryAxis fx:id="xAxis" label="Month" />
+				</xAxis>
+				<yAxis>
+					<NumberAxis fx:id="yAxis" label="Sessions" />
+				</yAxis>
+			</BarChart>
+		</SplitPane>
+
+	</VBox>
 </AnchorPane>
