+++ Client/src/client/ViewSubscriberController.java
@@ -1,25 +1,17 @@
 package client;
 
 import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import java.time.Month;
-import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import javafx.application.Platform;
-import javafx.beans.binding.Bindings;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.scene.chart.BarChart;
-import javafx.scene.chart.CategoryAxis;
-import javafx.scene.chart.NumberAxis;
-import javafx.scene.chart.XYChart;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.ButtonType;
@@ -33,6 +25,16 @@ import logic.Role;
 import logic.SendObject;
 import logic.subscriber;
 
+/**
+ * Controller for viewing subscribers and their parking session history in a
+ * client-server parking management system. This class provides functionality
+ * for filtering, sorting, and viewing subscriber data, as well as displaying
+ * parking session history for individual subscribers.
+ * <p>
+ * The controller interacts with the server to fetch subscriber and parking
+ * session data, and updates the UI components accordingly.
+ * </p>
+ */
 public class ViewSubscriberController extends Controller {
 
 	@FXML
@@ -64,39 +66,34 @@ public class ViewSubscriberController extends Controller {
 	protected final List<String> roles = Arrays.asList("All", "MANAGER", "SUBSCRIBER", "ATTENDANT");
 
 	protected Runnable backHandler;
-	
+
 	@FXML
-    private TextField subscriberIdField;
+	private TextField subscriberIdField;
 
-    @FXML
-    private Button viewHistoryButton;
+	@FXML
+	private Button viewHistoryButton;
 
-    @FXML
-    private TableView<Parkingsession> historyTable;
+	@FXML
+	private TableView<Parkingsession> historyTable;
 
-    @FXML
-    private TableColumn<Parkingsession, Integer> colSessionId;
-    @FXML
-    private TableColumn<Parkingsession, Integer> colSpotId;
-    @FXML
-    private TableColumn<Parkingsession, String> colInTime;
-    @FXML
-    private TableColumn<Parkingsession, String> colOutTime;
-    @FXML
-    private TableColumn<Parkingsession, String> colLate;
+	@FXML
+	private TableColumn<Parkingsession, Integer> colSessionId;
+	@FXML
+	private TableColumn<Parkingsession, Integer> colSpotId;
+	@FXML
+	private TableColumn<Parkingsession, String> colInTime;
+	@FXML
+	private TableColumn<Parkingsession, String> colOutTime;
+	@FXML
+	private TableColumn<Parkingsession, String> colLate;
 
-    @FXML
-    private BarChart<String, Number> barChart;
-    @FXML
-    private CategoryAxis xAxis;
-    @FXML
-    private NumberAxis yAxis;
-    
-    private boolean responseReceived = false;
-    private List<Parkingsession> historySessions;
+	private List<Parkingsession> historySessions;
 
-    private final HistoryController controller = new HistoryController();
-
+	/**
+	 * Initializes the UI components and sets up event handlers for buttons and
+	 * combo boxes. This method is automatically called after the FXML file is
+	 * loaded.
+	 */
 	@FXML
 	public void initialize() {
 		// Column setup
@@ -116,77 +113,56 @@ public class ViewSubscriberController extends Controller {
 		if (sortBySubscriberIdButton != null) {
 			sortBySubscriberIdButton.setOnAction(e -> sortBySubscriberId());
 		}
-		
+
 		// History table setup
 		colSessionId.setCellValueFactory(new PropertyValueFactory<>("sessionId"));
-        colSpotId.setCellValueFactory(new PropertyValueFactory<>("spotId"));
-        colInTime.setCellValueFactory(cellData -> javafx.beans.binding.Bindings.createStringBinding(() ->
-            new SimpleDateFormat("dd/MM/yyyy HH:mm").format(cellData.getValue().getInTime())));
-        colOutTime.setCellValueFactory(cellData -> javafx.beans.binding.Bindings.createStringBinding(() ->
-            new SimpleDateFormat("dd/MM/yyyy HH:mm").format(cellData.getValue().getOutTime())));
-        colLate.setCellValueFactory(cellData -> javafx.beans.binding.Bindings.createStringBinding(() ->
-            cellData.getValue().isLate() ? "Yes" : "No"));
-        viewHistoryButton.setOnAction(e -> fetchHistory());
+		colSpotId.setCellValueFactory(new PropertyValueFactory<>("spotId"));
+		colInTime.setCellValueFactory(cellData -> javafx.beans.binding.Bindings.createStringBinding(
+				() -> new SimpleDateFormat("dd/MM/yyyy HH:mm").format(cellData.getValue().getInTime())));
+		colOutTime.setCellValueFactory(cellData -> javafx.beans.binding.Bindings.createStringBinding(
+				() -> new SimpleDateFormat("dd/MM/yyyy HH:mm").format(cellData.getValue().getOutTime())));
+		colLate.setCellValueFactory(cellData -> javafx.beans.binding.Bindings
+				.createStringBinding(() -> cellData.getValue().isLate() ? "Yes" : "No"));
+		viewHistoryButton.setOnAction(e -> fetchHistory());
 	}
-	
+
+	/**
+	 * Fetches the parking session history for the subscriber with the entered ID.
+	 * Sends a request to the server to retrieve the history data.
+	 * 
+	 * @throws NumberFormatException if the entered subscriber ID is not a valid
+	 *                               number.
+	 */
 	private void fetchHistory() {
-        String idText = subscriberIdField.getText().trim();
-        if (idText.isEmpty()) {
-            showAlert("Please enter a Subscriber ID");
-            return;
-        }
+		String idText = subscriberIdField.getText().trim();
+		if (idText.isEmpty()) {
+			showAlert("Please enter a Subscriber ID");
+			return;
+		}
 
-        try {
-            int subscriberId = Integer.parseInt(idText);
-            responseReceived = false;
-            client.sendToServerSafely(new SendObject<Integer>("Get history", subscriberId));
+		try {
+			int subscriberId = Integer.parseInt(idText);
+			client.sendToServerSafely(new SendObject<Integer>("Get history", subscriberId));
+		} catch (NumberFormatException e) {
+			showAlert("Subscriber ID must be a number.");
+		}
+	}
 
-            new Thread(() -> {
-                try {
-                    Util.waitForServerResponse(10000, () -> responseReceived);
-                    List<Parkingsession> result = historySessions;
-                    if (result != null) {
-                        Platform.runLater(() -> {
-                            populateHistoryTable(result);
-                            populateBarChart(result);
-                        });
-                    } else {
-                        Platform.runLater(() -> showAlert("No session data received."));
-                    }
-                } catch (Exception ex) {
-                    Platform.runLater(() -> showAlert("Error: " + ex.getMessage()));
-                }
-            }).start();
-        } catch (NumberFormatException e) {
-            showAlert("Subscriber ID must be a number.");
-        }
-    }
+	/**
+	 * Populates the history table with the provided list of parking sessions.
+	 * 
+	 * @param sessions The list of parking sessions to display in the history table.
+	 */
+	private void populateHistoryTable(List<Parkingsession> sessions) {
+		ObservableList<Parkingsession> data = FXCollections.observableArrayList(sessions);
+		historyTable.setItems(data);
+	}
 
-    private void populateHistoryTable(List<Parkingsession> sessions) {
-        ObservableList<Parkingsession> data = FXCollections.observableArrayList(sessions);
-        historyTable.setItems(data);
-    }
-
-    private void populateBarChart(List<Parkingsession> sessions) {
-        int[] monthly = new int[12];
-        int currentYear = LocalDate.now().getYear();
-
-        for (Parkingsession s : sessions) {
-            LocalDate date = s.getInTime().toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
-            if (date.getYear() == currentYear) {
-                monthly[date.getMonthValue() - 1]++;
-            }
-        }
-
-        XYChart.Series<String, Number> series = new XYChart.Series<>();
-        for (int i = 0; i < 12; i++) {
-            series.getData().add(new XYChart.Data<>(Month.of(i + 1).name(), monthly[i]));
-        }
-        barChart.getData().clear();
-        barChart.getData().add(series);
-    }
-
-	/** Can be reused in subclass */
+	/**
+	 * Can be reused in subclass. 
+	 * Filters the subscriber list based on the selected
+	 * role in the combo box. Updates the table view with the filtered list.
+	 */
 	protected void filterSubscribers() {
 		String selectedRole = roleComboBox.getValue();
 
@@ -199,44 +175,72 @@ public class ViewSubscriberController extends Controller {
 		subscriberTable.getItems().setAll(filteredSubscribers);
 	}
 
+    /**
+     * Sorts the subscriber table by subscriber ID in ascending order.
+     */
 	protected void sortBySubscriberId() {
 		colSubscriberId.setSortType(TableColumn.SortType.ASCENDING);
 		subscriberTable.getSortOrder().setAll(colSubscriberId);
 		subscriberTable.sort();
 	}
 
+    /**
+     * Sets the list of subscribers and updates the table view.
+     * 
+     * @param subscribers The list of subscribers to display in the table view.
+     */
 	public void setSubscribers(List<subscriber> subscribers) {
 		this.allSubscribers = subscribers;
 		filterSubscribers();
 	}
 
-	@Override
-	public void handleServerMessage(Object msg) {
-	    if (msg instanceof SendObject<?>) {
-	        SendObject<?> so = (SendObject<?>) msg;
-
-	        // Handle subscriber list updates
-	        if (so.getObj() instanceof List<?>) {
-	            List<?> updated = (List<?>) so.getObj();
-	            if (!updated.isEmpty() && updated.get(0) instanceof subscriber) {
-	                Platform.runLater(() -> setSubscribers((List<subscriber>) updated));
-	            }
-	        }
-
-	        // Handle history data per specific subscriber
-	        if ("Parkingsession list of subscriber".equals(so.getObjectMessage())) {
-	            Object obj = so.getObj();
-	            if (obj instanceof List<?>) {
-	                List<?> list = (List<?>) obj;
-	                if (!list.isEmpty() && list.get(0) instanceof Parkingsession) {
-	                    historySessions = (List<Parkingsession>) list;
-	                    responseReceived = true;
-	                }
-	            }
-	        }
-	    }
+    /**
+     * Sets the list of parking session history and updates the history table.
+     * 
+     * @param history The list of parking sessions to display in the history table.
+     */
+	public void setHistorySessions(List<Parkingsession> history) {
+		if (history != null) {
+			this.historySessions = history;
+		} else {
+			this.historySessions = new ArrayList<Parkingsession>();
+		}
+		populateHistoryTable(history);
 	}
 
+    /**
+     * Handles server messages and updates the UI components based on the received data.
+     * 
+     * @param msg The message received from the server.
+     */
+	@Override
+	public void handleServerMessage(Object msg) {
+		if (msg instanceof SendObject<?>) {
+			SendObject<?> so = (SendObject<?>) msg;
+
+			// Handle subscriber list updates
+			if (so.getObj() instanceof List<?>) {
+				List<?> updated = (List<?>) so.getObj();
+				if (!updated.isEmpty() && updated.get(0) instanceof subscriber) {
+					Platform.runLater(() -> setSubscribers((List<subscriber>) updated));
+				}
+			}
+
+			// Handle history data per specific subscriber
+			if ("Parkingsession list of subscriber".equals(so.getObjectMessage())) {
+				Object obj = so.getObj();
+				if (obj instanceof List<?>) {
+					List<?> list = (List<?>) obj;
+					if (!list.isEmpty() && list.get(0) instanceof Parkingsession) {
+						setHistorySessions((List<Parkingsession>) list);
+					} else {
+						// empty the table and chart if no sessions found
+						setHistorySessions(new ArrayList<>());
+					}
+				}
+			}
+		}
+	}
 
 	protected void showAlert(String msg) {
 		Alert alert = new Alert(Alert.AlertType.INFORMATION, msg, ButtonType.OK);
